Flatten AMP components structure & fix imports
