+++ dotcom-rendering/src/web/components/Branding.amp.tsx
@@ -1,11 +1,11 @@
 import { css } from '@emotion/react';
 import { textSans } from '@guardian/source-foundations';
 import React from 'react';
-import { neutralBorder, pillarPalette_DO_NOT_USE } from '../../../lib/pillars';
-import type { Branding as BrandingType } from '../../../types/branding';
-import type { CommercialProperties } from '../../../types/commercial';
-import { isEditionId } from '../../lib/edition';
-import { regionClasses } from '../../lib/region-classes.amp';
+import { neutralBorder, pillarPalette_DO_NOT_USE } from '../../lib/pillars';
+import type { Branding as BrandingType } from '../../types/branding';
+import type { CommercialProperties } from '../../types/commercial';
+import { isEditionId } from '../lib/edition';
+import { regionClasses } from '../lib/region-classes.amp';
 
 const LinkStyle = (pillar: ArticleTheme) => css`
 	a {
