+++ dotcom-rendering/src/web/components/ContentAtomBlockComponent.amp.tsx
