+++ dotcom-rendering/src/web/components/Elements.amp.tsx
@@ -3,26 +3,26 @@ import { NotRenderableInDCR } from '../../lib/errors/not-renderable-in-dcr';
 import type { Switches } from '../../types/config';
 import type { FEElement } from '../../types/content';
 import type { TagType } from '../../types/tag';
-import { AudioAtomBlockComponent } from '../../web/components/AMPElements/AudioAtomBlockComponent.amp';
-import { CommentBlockComponent } from '../../web/components/AMPElements/CommentBlockComponent.amp';
-import { ContentAtomBlockComponent } from '../../web/components/AMPElements/ContentAtomBlockComponent.amp';
-import { DisclaimerBlockComponent } from '../../web/components/AMPElements/DisclaimerBlockComponent.amp';
-import { EmbedBlockComponent } from '../../web/components/AMPElements/EmbedBlockComponent.amp';
-import { GuVideoBlockComponent } from '../../web/components/AMPElements/GuVideoBlockComponent.amp';
-import { ImageBlockComponent } from '../../web/components/AMPElements/ImageBlockComponent.amp';
-import { InteractiveAtomBlockComponent } from '../../web/components/AMPElements/InteractiveAtomBlockComponent.amp';
-import { InteractiveBlockComponent } from '../../web/components/AMPElements/InteractiveBlockComponent.amp';
-import { PullquoteBlockComponent } from '../../web/components/AMPElements/PullquoteBlockComponent.amp';
-import { RichLinkBlockComponent } from '../../web/components/AMPElements/RichLinkBlockComponent.amp';
-import { SoundcloudBlockComponent } from '../../web/components/AMPElements/SoundcloudBlockComponent.amp';
-import { SubheadingBlockComponent } from '../../web/components/AMPElements/SubheadingBlockComponent.amp';
-import { TextBlockComponent } from '../../web/components/AMPElements/TextBlockComponent.amp';
-import { TimelineBlockComponent } from '../../web/components/AMPElements/TimelineBlockComponent.amp';
-import { TwitterBlockComponent } from '../../web/components/AMPElements/TwitterBlockComponent.amp';
-import { VideoVimeoBlockComponent } from '../../web/components/AMPElements/VideoVimeoBlockComponent.amp';
-import { VideoYoutubeBlockComponent } from '../../web/components/AMPElements/VideoYoutubeBlockComponent.amp';
-import { YoutubeBlockComponent } from '../../web/components/AMPElements/YoutubeBlockComponent.amp';
+import { AudioAtomBlockComponent } from './AudioAtomBlockComponent.amp';
+import { CommentBlockComponent } from './CommentBlockComponent.amp';
+import { ContentAtomBlockComponent } from './ContentAtomBlockComponent.amp';
+import { DisclaimerBlockComponent } from './DisclaimerBlockComponent.amp';
+import { EmbedBlockComponent } from './EmbedBlockComponent.amp';
 import { Expandable } from './Expandable.amp';
+import { GuVideoBlockComponent } from './GuVideoBlockComponent.amp';
+import { ImageBlockComponent } from './ImageBlockComponent.amp';
+import { InteractiveAtomBlockComponent } from './InteractiveAtomBlockComponent.amp';
+import { InteractiveBlockComponent } from './InteractiveBlockComponent.amp';
+import { PullquoteBlockComponent } from './PullquoteBlockComponent.amp';
+import { RichLinkBlockComponent } from './RichLinkBlockComponent.amp';
+import { SoundcloudBlockComponent } from './SoundcloudBlockComponent.amp';
+import { SubheadingBlockComponent } from './SubheadingBlockComponent.amp';
+import { TextBlockComponent } from './TextBlockComponent.amp';
+import { TimelineBlockComponent } from './TimelineBlockComponent.amp';
+import { TwitterBlockComponent } from './TwitterBlockComponent.amp';
+import { VideoVimeoBlockComponent } from './VideoVimeoBlockComponent.amp';
+import { VideoYoutubeBlockComponent } from './VideoYoutubeBlockComponent.amp';
+import { YoutubeBlockComponent } from './YoutubeBlockComponent.amp';
 
 /**
  * Elements we support on AMP -
