+++ dotcom-rendering/src/web/components/EmbedBlockComponent.amp.tsx
@@ -1,5 +1,5 @@
-import { NotRenderableInDCR } from '../../../lib/errors/not-renderable-in-dcr';
-import type { EmbedBlockElement } from '../../../types/content';
+import { NotRenderableInDCR } from '../../lib/errors/not-renderable-in-dcr';
+import type { EmbedBlockElement } from '../../types/content';
 
 type Props = {
 	element: EmbedBlockElement;
