+++ dotcom-rendering/src/web/components/ImageBlockComponent.amp.tsx
@@ -1,9 +1,9 @@
 import { css } from '@emotion/react';
 import { text, textSans } from '@guardian/source-foundations';
-import { pillarPalette_DO_NOT_USE } from '../../../lib/pillars';
-import TriangleIcon from '../../../static/icons/triangle.svg';
-import type { ImageBlockElement } from '../../../types/content';
-import { bestFitImage, heightEstimate } from '../../lib/image-fit.amp';
+import { pillarPalette_DO_NOT_USE } from '../../lib/pillars';
+import TriangleIcon from '../../static/icons/triangle.svg';
+import type { ImageBlockElement } from '../../types/content';
+import { bestFitImage, heightEstimate } from '../lib/image-fit.amp';
 
 const figureStyle = css`
 	margin-top: 16px;
