+++ dotcom-rendering/src/web/components/PullquoteBlockComponent.amp.tsx
@@ -1,7 +1,7 @@
 import { css } from '@emotion/react';
 import { body, neutral } from '@guardian/source-foundations';
-import { pillarPalette_DO_NOT_USE } from '../../../lib/pillars';
-import Quote from '../../../static/icons/quote.svg';
+import { pillarPalette_DO_NOT_USE } from '../../lib/pillars';
+import Quote from '../../static/icons/quote.svg';
 
 const styles = (pillar: ArticleTheme) => css`
 	background-color: ${neutral[97]};
