+++ dotcom-rendering/src/web/components/TopMeta.amp.tsx
@@ -1,5 +1,5 @@
 import { ArticleDesign, ArticleSpecial } from '@guardian/libs';
-import type { AMPArticleModel } from '../../../../types/ampArticleModel';
+import type { AMPArticleModel } from '../../types/ampArticleModel';
 import { TopMetaAnalysis } from './TopMetaAnalysis.amp';
 import { TopMetaNews } from './TopMetaNews.amp';
 import { TopMetaOpinion } from './TopMetaOpinion.amp';
