+++ dotcom-rendering/src/web/components/TopMetaOpinion.amp.tsx
@@ -1,13 +1,13 @@
 import { css } from '@emotion/react';
 import { headline, neutral } from '@guardian/source-foundations';
-import type { AMPArticleModel } from '../../../../types/ampArticleModel';
-import { MainMedia } from '../../../amp/components/MainMedia.amp';
-import { getAgeWarning } from '../../../lib/age-warning';
-import { getSoleContributor } from '../../../lib/byline';
-import { pillarPalette_DO_NOT_USE } from '../../../lib/pillars';
-import { getSharingUrls } from '../../../lib/sharing-urls';
+import { getAgeWarning } from '../../lib/age-warning';
+import { getSoleContributor } from '../../lib/byline';
+import { pillarPalette_DO_NOT_USE } from '../../lib/pillars';
+import { getSharingUrls } from '../../lib/sharing-urls';
+import type { AMPArticleModel } from '../../types/ampArticleModel';
 import { Branding, BrandingRegionContainer } from './Branding.amp';
 import { Byline } from './Byline.amp';
+import { MainMedia } from './MainMedia.amp';
 import { SeriesLink } from './SeriesLink.amp';
 import { Standfirst } from './Standfirst.amp';
 import { TopMetaExtras } from './TopMetaExtras.amp';
