+++ dotcom-rendering/src/web/components/TopMetaPaidContent.amp.tsx
@@ -1,13 +1,13 @@
 import { css } from '@emotion/react';
 import { body, neutral, text, textSans } from '@guardian/source-foundations';
-import type { AMPArticleModel } from '../../../../types/ampArticleModel';
-import { MainMedia } from '../../../amp/components/MainMedia.amp';
-import { getAgeWarning } from '../../../lib/age-warning';
-import { getSoleContributor } from '../../../lib/byline';
-import { getSharingUrls } from '../../../lib/sharing-urls';
-import type { Branding } from '../../../types/branding';
+import { getAgeWarning } from '../../lib/age-warning';
+import { getSoleContributor } from '../../lib/byline';
+import { getSharingUrls } from '../../lib/sharing-urls';
+import type { AMPArticleModel } from '../../types/ampArticleModel';
+import type { Branding } from '../../types/branding';
 import { BrandingRegionContainer } from './Branding.amp';
 import { Byline } from './Byline.amp';
+import { MainMedia } from './MainMedia.amp';
 import { PaidForBand } from './PaidForBand.amp';
 import { Standfirst } from './Standfirst.amp';
 import { TopMetaExtras } from './TopMetaExtras.amp';
