+++ dotcom-rendering/src/web/components/VideoVimeoBlockComponent.amp.tsx
@@ -1,6 +1,6 @@
-import type { VideoVimeoBlockElement } from '../../../types/content';
-import { getIdFromUrl } from '../../lib/get-video-id.amp';
-import { Caption } from '../../../amp/components/Caption.amp';
+import type { VideoVimeoBlockElement } from '../../types/content';
+import { getIdFromUrl } from '../lib/get-video-id.amp';
+import { Caption } from './Caption.amp';
 
 type Props = {
 	element: VideoVimeoBlockElement;
