+++ dotcom-rendering/src/web/server/render.article.web.tsx
@@ -3,7 +3,6 @@ import {
 	BUILD_VARIANT,
 	dcrJavascriptBundle,
 } from '../../../scripts/webpack/bundles';
-import { isAmpSupported } from '../../amp/components/Elements.amp';
 import { buildAdTargeting } from '../../lib/ad-targeting';
 import {
 	ASSET_ORIGIN,
@@ -18,6 +17,7 @@ import type { FEElement } from '../../types/content';
 import type { FEArticleType } from '../../types/frontend';
 import type { TagType } from '../../types/tag';
 import { ArticlePage } from '../components/ArticlePage';
+import { isAmpSupported } from '../components/Elements.amp';
 import { KeyEventsContainer } from '../components/KeyEventsContainer';
 import { decideFormat } from '../lib/decideFormat';
 import { decideTheme } from '../lib/decideTheme';
